backend/requirements.txt
python-dotenv
typer
fastapi
pydantic
elevenlabs
srt
torch
torchaudio
openai-whisper
beautifulsoup4
markdownify
soundfile
openai
pandas
transformers
lmnt
python-youtube
arxiv
ffmpeg
joblib